+++ Assets/Scripts/WaveSynth/WaveOutputs/WaveGenerators/SquareGenerator.cs
@@ -2,7 +2,7 @@
 {
     public class SquareGenerator : FunctionGenerator
     {
-        protected override float SampleFunction(float phase)
+        protected override float SampleFunction(double phase)
         {
             return phase < 0.5f ? -1 : 1;
         }
